+++ src/core/themes/darkTheme.json
